==> R/03-data_splitting.R <==
library(group18package)
# takes in heart_data.csv generated by 02

# produces 2 csv files inside data/modelling.
# one that is the testing data and another that is the training data

library(tidymodels)

source('R/joining_data.R')

set.seed(1)

# reading file created in 02
heart_data <- clean_csv('data/processed/heart_data.csv')

# initial splitting of the data 
heart_split<-initial_split(heart_data, prop=0.75, strata=diagnosis_f)
heart_training<-training(heart_split)
heart_testing<-testing(heart_split)

# Saving initial test data (this will be over written after forward selection is run)
write_csv(heart_training, 'data/modelling/training_split.csv')
write_csv(heart_testing, 'data/modelling/testing_split.csv')
