==> R/07.0-build_classification_model.R <==
library(group18package)
# This script builds a table of the classification model
# takes in data generated by 03 script

# produces a table from the training split

source('R/classification_model.R')
source('R/joining_data.R')

# build the table and classifies it
heart_data_accuracies <- classifier(clean_csv('data/modelling/training_split_new.csv'))

# writes png to figures folder
# write_table(heart_data_accuracies, '../figures/classification_model.csv')
write_csv(heart_data_accuracies, "data/modelling/heart_data_accuracies.csv")
