==> R/07.1-visualize_classification_model.R <==
library(group18package)
# # This script takes in the table and builds a visualization
# takes in data generated by 07.0 script

# produces a knn visualization as a png
library(tidyverse)

# Reading in data from 07.0
data <- read_csv("data/modelling/heart_data_accuracies.csv")

# builds a knn visualization from the classification model from figures folder
knn_visual <- knn_visualization(data)

# saves the knn visualization as a png in the figures folder
ggsave('figures/classification_model_visualization.png', knn_visual)
